Update ServiceDelegate.js
Lose `characteristics` property in favour of `addCharacteristic()`.  Problem was that `charactertistic` was accessed from the `ServiceDelegete` supertype constructor, before the subtype was fully initialised (causing `this` to be `undefined`).  `addCharacteristic()` takes a single parameter, `params`, which has the same structure as an element of the `charactertistics` array.
Also added `params.didSet`, a function to be called as handler for the `didSet` event.